requirements.txt
gevent==1.0.1
greenlet==0.4.5

PyYAML==3.11
